+++ src/App.css
@@ -8,6 +8,7 @@
   box-sizing: border-box;
   font-family: "DM Serif Text", serif;
 }
+
 a {
   text-decoration: none;
 }
@@ -20,6 +21,7 @@ a {
   --color: #f5fcffda;
   --btn: #655dbb63;
 }
+
 .wrap {
   position: relative;
 }
@@ -27,10 +29,12 @@ a {
 .l-wrap-outer {
   padding: 0 60px;
 }
+
 .l-wrap-inner {
   max-width: 1171px;
   margin: 0 auto;
 }
+
 .bg {
   background-image: url(./img/bg.jpg);
   background-position: center center;
@@ -62,10 +66,11 @@ ul {
 }
 
 /* blog */
-#blog{
-  scroll-margin-top: 100px;
-}
 
+#blog {
+  scroll-margin-top: 100px;
+  margin-top: 80px;
+}
 
 .blog_wrap {
   display: flex;
@@ -76,9 +81,11 @@ ul {
 }
 
 /* contact */
-#contact{
+#contact {
   scroll-margin-top: 50px;
+  margin-top: 80px;
 }
+
 .content {
   margin: 0 auto;
   width: 650px;
@@ -140,7 +147,6 @@ hr {
 }
 
 
-
 textarea {
   width: 600px;
   height: 100px;
@@ -150,15 +156,19 @@ textarea {
   padding-top: 10px;
   padding-left: 10px;
 }
+
 /* blog detail */
- #blog-detail{
+#blog-detail {
   scroll-margin-top: 85px;
- }
+}
 
 /* about us */
-#about{
-  scroll-margin-top: 100px;
+
+#about {
+  margin-top: 120px;
+  /* scroll-margin-top: 150px; */
 }
+
 .about-container {
   margin-top: 80px;
 }
@@ -178,7 +188,7 @@ textarea {
 }
 
 /* Login page */
-.login{
+.login {
   scroll-margin-top: 0px;
 }
 
@@ -200,6 +210,7 @@ ol,
 ul {
   list-style: none;
 }
+
 footer {
   padding: 40px 0;
   background-color: #050542;
